refactor: move Endpoint and TLS helpers out of router_factory.rs
router_factory.rs hosted two residents unrelated to the factory
concern. Endpoint/EndpointHandler move to axum_factory, where the
type's consumers live and its into_router conversion belongs; the
lib.rs re-export keeps the public apollo_router::Endpoint path
unchanged. The TlsClient certificate-store impl and its helpers move
to configuration/mod.rs beside TlsClient and the existing certificate
loaders; the moved private loader is named load_root_certs to avoid
colliding with configuration::load_certs, whose error mapping
differs. router_factory.rs now holds only the two traits,
YamlRouterFactory, and their tests.
