+++ apollo-router/src/axum_factory/axum_http_server_factory.rs
@@ -43,6 +43,7 @@ use super::utils::ConnectionRouterService;
 use super::utils::PropagatingMakeSpan;
 use super::utils::connection_router_service;
 use crate::Context;
+use crate::axum_factory::Endpoint;
 use crate::axum_factory::compression::Compressor;
 use crate::axum_factory::listeners::get_extra_listeners;
 use crate::axum_factory::listeners::serve_router_on_listen_addr;
@@ -57,7 +58,6 @@ use crate::plugins::license_enforcement::layer::LicenseLayer;
 use crate::plugins::telemetry::config_new::router::instruments::ResponseBodySizeRecording;
 use crate::plugins::telemetry::config_new::router::instruments::ResponseBodySizeRecordingStream;
 use crate::router::ApolloRouterError;
-use crate::router_factory::Endpoint;
 use crate::router_factory::RouterFactory;
 use crate::services::router;
 use crate::uplink::license_enforcement::LicenseState;
