+++ apollo-router/src/axum_factory/mod.rs
@@ -2,6 +2,7 @@
 pub(crate) mod axum_http_server_factory;
 pub(crate) mod compression;
 pub(crate) mod connection_handle;
+mod endpoint;
 mod listeners;
 #[cfg(test)]
 pub(crate) mod tests;
@@ -13,6 +14,7 @@ use std::sync::OnceLock;
 use axum::Router;
 pub(crate) use axum_http_server_factory::AxumHttpServerFactory;
 pub(crate) use axum_http_server_factory::CanceledRequest;
+pub use endpoint::Endpoint;
 pub(crate) use listeners::ListenAddrAndRouter;
 
 static ENDPOINT_CALLBACK: OnceLock<Arc<dyn Fn(Router) -> Router + Send + Sync>> = OnceLock::new();
