+++ apollo-router/src/axum_factory/tests.rs
@@ -65,6 +65,7 @@ use crate::Configuration;
 use crate::ListenAddr;
 use crate::TestHarness;
 use crate::assert_response_eq_ignoring_error_id;
+use crate::axum_factory::Endpoint;
 use crate::configuration::Homepage;
 use crate::configuration::Sandbox;
 use crate::configuration::Supergraph;
@@ -75,7 +76,6 @@ use crate::http_server_factory::HttpServerFactory;
 use crate::http_server_factory::HttpServerHandle;
 use crate::json_ext::Path;
 use crate::plugins::healthcheck::Config as HealthCheck;
-use crate::router_factory::Endpoint;
 use crate::router_factory::RouterFactory;
 use crate::services::MULTIPART_DEFER_ACCEPT;
 use crate::services::MULTIPART_DEFER_CONTENT_TYPE;
